ref(autopilot): Split tasks.py into separate detector modules
Split the monolithic tasks.py file into a tasks/ package with separate
modules for each detector, improving code organization and maintainability.

- common.py: Shared utilities (AutopilotDetectorName, create_instrumentation_issue)
- sdk_update.py: SDK update detector
- missing_sdk_integration.py: Missing SDK integration detector
- trace_instrumentation.py: Trace instrumentation detector

Also split tests into corresponding test files.
