+++ src/sentry/autopilot/tasks/__init__.py
@@ -1,0 +1,9 @@
+from sentry.autopilot.tasks.missing_sdk_integration import run_missing_sdk_integration_detector
+from sentry.autopilot.tasks.sdk_update import run_sdk_update_detector
+from sentry.autopilot.tasks.trace_instrumentation import run_trace_instrumentation_detector
+
+__all__ = [
+    "run_missing_sdk_integration_detector",
+    "run_sdk_update_detector",
+    "run_trace_instrumentation_detector",
+]
